requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.7
langchain-community==0.3.7
langchain-google-genai==2.0.4
ollama==0.3.3
chromadb==0.5.18
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
pypdf==4.3.1
python-docx==1.1.2
 

